+++ packages/element/src/index.ts
@@ -22,6 +22,7 @@ export * from '@availity/mui-page-header';
 export * from '@availity/mui-pagination';
 export * from '@availity/mui-paper';
 export * from '@availity/mui-popover';
+export * from '@availity/mui-popper';
 export * from '@availity/mui-progress';
 export * from '@availity/mui-table';
 export * from '@availity/mui-tabs';