+++ packages/avatar/CHANGELOG.md
@@ -6,7 +6,8 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ### Dependency Updates
 
-* `design-tokens` updated to version `0.12.1`
+- `design-tokens` updated to version `0.12.1`
+
 ## [0.1.9](https://github.com/Availity/element/compare/@availity/mui-avatar@0.1.8...@availity/mui-avatar@0.1.9) (2023-11-16)
 
 ### Dependency Updates