+++ packages/theme/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.31.2](https://github.com/Availity/element/compare/@availity/theme@0.31.1...@availity/theme@0.31.2) (2023-12-14)
+
+
+### Bug Fixes
+
+* **theme:** legacy container maxWidth ([669fec5](https://github.com/Availity/element/commit/669fec5f81cb161423b649bc705248aa0700985b))
+
+## [0.31.1](https://github.com/Availity/element/compare/@availity/theme@0.31.0...@availity/theme@0.31.1) (2023-12-14)
+
+## [0.31.0](https://github.com/Availity/element/compare/@availity/theme@0.30.4...@availity/theme@0.31.0) (2023-12-14)
+
+
+### Features
+
+* add form story ([11d1f6c](https://github.com/Availity/element/commit/11d1f6c6191540bf02ff42d4056ac553eecc8c73))
+
+## [0.30.4](https://github.com/Availity/element/compare/@availity/theme@0.30.3...@availity/theme@0.30.4) (2023-12-13)
+
 ## [0.30.3](https://github.com/Availity/element/compare/@availity/theme@0.30.2...@availity/theme@0.30.3) (2023-12-07)
 
 ## [0.30.2](https://github.com/Availity/element/compare/@availity/theme@0.30.1...@availity/theme@0.30.2) (2023-12-06)
@@ -20,29 +38,27 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ## [0.29.0](https://github.com/Availity/element/compare/@availity/theme@0.28.1...@availity/theme@0.29.0) (2023-11-28)
 
-
 ### Features
 
-* **theme:** light theme autocomplete styles ([66d9645](https://github.com/Availity/element/commit/66d9645673d7054d77dbaabbbc9ba77b78aa5a52))
+- **theme:** light theme autocomplete styles ([66d9645](https://github.com/Availity/element/commit/66d9645673d7054d77dbaabbbc9ba77b78aa5a52))
 
 ## [0.28.1](https://github.com/Availity/element/compare/@availity/theme@0.28.0...@availity/theme@0.28.1) (2023-11-27)
 
 ### Dependency Updates
 
-* `design-tokens` updated to version `0.12.1`
-## [0.28.0](https://github.com/Availity/element/compare/@availity/theme@0.27.0...@availity/theme@0.28.0) (2023-11-22)
+- `design-tokens` updated to version `0.12.1`
 
+## [0.28.0](https://github.com/Availity/element/compare/@availity/theme@0.27.0...@availity/theme@0.28.0) (2023-11-22)
 
 ### Features
 
-* **theme:** legacy autocomplete styles ([afce32e](https://github.com/Availity/element/commit/afce32e8f0c3d620bbc94ec788499e53af5c68bd))
+- **theme:** legacy autocomplete styles ([afce32e](https://github.com/Availity/element/commit/afce32e8f0c3d620bbc94ec788499e53af5c68bd))
 
 ## [0.27.0](https://github.com/Availity/element/compare/@availity/theme@0.26.0...@availity/theme@0.27.0) (2023-11-21)
 
-
 ### Features
 
-* **mui-feedback:** initial commit ([2362d7f](https://github.com/Availity/element/commit/2362d7f892274f6260d4814fd8b19b85d8e4fe70))
+- **mui-feedback:** initial commit ([2362d7f](https://github.com/Availity/element/commit/2362d7f892274f6260d4814fd8b19b85d8e4fe70))
 
 ## [0.26.0](https://github.com/Availity/element/compare/@availity/theme@0.25.1...@availity/theme@0.26.0) (2023-11-16)